+++ Domain/Transports/TransportBlank.cs
@@ -1,4 +1,4 @@
 ﻿namespace Domain.Transports;
 
 public record TransportBlank(Boolean CanBeRented, String TransportType, String Model, String Color, String Identifier, String? Description, 
-    Decimal Latitude, Decimal Longitude, Decimal? MinutePrice, Decimal? DayPrice);
+    Double Latitude, Double Longitude, Decimal? MinutePrice, Decimal? DayPrice);
